% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorCells.R
\name{anchor.cell}
\alias{anchor.cell}
\title{Classify Raster Cells Based on Anchor Cells}
\usage{
anchor.cell(
  attTbl,
  rstack,
  anchor,
  class,
  classVector = NULL,
  class2cell = TRUE,
  class2nbs = TRUE,
  overwrite_class = FALSE,
  plot = TRUE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{rstack}{\code{RasterStack} or \code{RasterLayer} object used to compute
the attribute table (see \code{\link{attTbl}}).}

\item{anchor}{integer vector, contains the cell numbers to be considered as
anchor points.}

\item{class}{numeric, the class to attribute to cells meeting contiguity
conditions.}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}

\item{class2cell}{logic, attribute the new classification, defined by the
argument \code{class}, to \code{anchor} cells.}

\item{class2nbs}{logic, attribute the new classification, defined by the
argument \code{class}, to the cells adjacent to \code{anchor} cells.}

\item{overwrite_class}{logic, reclassify cells that have already been
classified.}

\item{plot}{logic, plot \code{classVector}.}

\item{writeRaster}{filename, if a raster name is provided save
\code{classVector} in a raster file.}

\item{overWrite}{logic, if the raster names already exist, the existing file
is overwritten.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function. If no \code{classVector} was provided, the function return a
new classification vector.
}
\description{
Return a classification vector based on anchor cells.
}
\details{
Classify cells of the \code{rstack} argument based on the anchor
cells provided by \code{anchor}. Return a classification vector that can be
indexed into \code{rstack} using the \code{Cell} column of the attribute
table.
}
\seealso{
\code{\link[=anchor.seed]{anchor.seed()}}, \code{\link[=anchor.svo]{anchor.svo()}}, \code{\link[=attTbl]{attTbl()}}
}
