% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorCells.R
\name{anchor.svo}
\alias{anchor.svo}
\title{Anchor cells from spatial vector objects}
\usage{
anchor.svo(
  rstack,
  spatial_vector_name,
  only_NAs = FALSE,
  fill_NAs = FALSE,
  plot = TRUE,
  saveRDS = NULL,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{spatial_vector_name}{OGR data source name.}

\item{only_NAs}{logic, cell numbers extracted only at the locations of a
spatial vector object that are not complete cases (i.e. have some missing
values in some of the layers of the \code{Raster*} object).}

\item{fill_NAs}{logic, cell numbers extracted also at locations contiguous to
those of the spatial vector object that are not complete cases (i.e. have
some missing values in some of the layers of the \code{Raster*} object).}

\item{plot}{logic, plot anchor points.}

\item{saveRDS}{filename, if a file name is provided save the anchor cell
vector as an RDS file.}

\item{writeRaster}{filename, if a raster name is provided save the anchor
cell vector as a raster file.}

\item{overWrite}{logic, if RDS and raster names already exist, existing files
are overwritten.}
}
\value{
Numeric vector of raster cell numbers.
}
\description{
Returns a vector of raster cell numbers extracted at the locations of a
spatial vector object.
}
\details{
When the arguments \code{only_NA} and \code{fill_NAs} are FALSE the
output is equivalent to the output of the function \code{raster::extract()}
over a raster whose values are cell numbers.
}
