% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInputs.R
\name{attTbl}
\alias{attTbl}
\title{Raster Object to Data Frame}
\usage{
attTbl(rstack, var_names = NULL)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{var_names}{character vector, re-name the variables of the
\code{Raster*} object in the \code{data.frame}.}
}
\value{
data.frame
}
\description{
Returns the attribute table (\code{data.frame}) used by other
\code{scapesClassification} functions. Each row corresponds to a cell on the
\code{Raster*} object.
}
\details{
Returns a \code{data.frame} with a number of columns equal to the
number of layers in \code{rstack} plus one and with a maximum number of
rows equal to the number of cells on \code{rstack}. Its first column, named
\code{'Cell'}, refers to positions on the \code{Raster*} object with
complete cases, i.e., raster cells having a value for every layer in the
stack.
}
\note{
The attribute table contains only complete cases, i.e., raster cells
having a value for every layer in the stack.
}
\examples{
## r1, raster with cell numbers as values
r1 <- raster::raster(matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE))

## r2, raster with missing values in cell numbers 1 and 2
r2 <- raster::raster(matrix(1, nrow = 3, ncol = 4, byrow = TRUE))
r2[c(1,2)] <- NA

attTbl(raster::stack(r1,r2))
attTbl(raster::stack(r1,r2), var_names = c("var_r1", "var_r2"))
}
