% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classContiguity.R
\name{classify.all}
\alias{classify.all}
\title{Classify All Unclassified Cells}
\usage{
classify.all(attTbl, ngbList, rNumb = FALSE, classVector)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{ngbList}{list, it has to contain the list of 8-neighbors of each cell
in \code{attTbl$Cell} (see \code{\link{ngbList}}).}

\item{rNumb}{logic, \code{ngbList} contain the neighbors index position
in the attribute table (see \code{\link{ngbList}}).}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function.
}
\description{
Classify all cells in \code{classVector} that have not yet been classified
based on contiguity conditions.
}
\details{
For each unclassified cell its 8-cell neighborhood is considered.
Among neighbors, the class with the highest number of members is assigned
to the unclassified cell. If two or more classes have the same number of
members then one of these classes is assigned randomly to the unclassified
cell. If none of the neighbors have been classified then no class is
assigned to the unclassified cell.
}
\seealso{
\code{\link[=attTbl]{attTbl()}}, \code{\link[=ngbList]{ngbList()}} and contiguity functions in \code{\link[=conditions]{conditions()}}
}
