% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionFunctions.R
\name{cond.4.all}
\alias{cond.4.all}
\title{Test Conditions for All Cells}
\usage{
cond.4.all(
  attTbl,
  conditions,
  classVector = NULL,
  class,
  overwrite_class = FALSE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{conditions}{character string, the conditions a cell have to meet to be
classified as indicated by the argument \code{class}. Absolute conditions
can be used (see \code{\link{conditions}}).}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}

\item{class}{numeric, the numeric class to attribute to the cells meeting
conditions.}

\item{overwrite_class}{logic, reclassify cells that have already been
classified.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function. If no \code{classVector} was provided, the function return a
new classification vector.
}
\description{
Evaluate conditions for cells that have not already been classified and
classify them if conditions are true.
}
\details{
The function evaluates the conditions of the argument
\code{\link{conditions}} for all cells that have not already been
classified. The cells for which \code{\link{conditions}} are \code{TRUE}
are classified as indicted by the argument \code{class}.
}
\seealso{
\code{\link[=conditions]{conditions()}}, \code{\link[=attTbl]{attTbl()}}
}
