% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cond.parse}
\alias{cond.parse}
\title{Parse conditions}
\usage{
cond.parse(names_attTbl, cond)
}
\arguments{
\item{names_attTbl}{character vector, the column (i.e. variable) names of the
attribute table returned by the function \code{\link{attTbl}}.}

\item{cond}{character string, the condition string used by the \code{cond.*}
functions to classify raster cells (see \code{\link{conditions}}).}
}
\value{
The function returns a two-element list. The first element contains
the parsed conditions to be evaluated by the \code{cond.*} functions. The
second element defines the condition type each variable refers to.
}
\description{
Parse the condition string so that it can be evaluated by the \verb{cond.*}
functions. Intended for internal use only.
}
\seealso{
\code{\link[=cond.4.all]{cond.4.all()}}, \code{\link[=cond.4.nofn]{cond.4.nofn()}}, \code{\link[=anchor.seed]{anchor.seed()}}, \code{\link[=cond.reclass]{cond.reclass()}},
\code{\link[=conditions]{conditions()}}
}
