% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aidFunctions.R
\name{cv.2.rast}
\alias{cv.2.rast}
\title{Class Vector To Raster}
\usage{
cv.2.rast(
  r,
  index,
  classVector,
  plot = TRUE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{r}{The \code{Raster*} object on which classification is based (see
\code{\link{attTbl}}).}

\item{index}{The cell numbers of complete cases, i.e., raster cells having a
value for every layer in the stack (see \code{\link{attTbl}}).}

\item{classVector}{numeric vector, defines the raster cell that have already
been classified.}

\item{plot}{logic, plot anchor \code{classVector}.}

\item{writeRaster}{filename, if a raster name is provided save the
\code{classVector} as a raster file.}

\item{overWrite}{logic, overwrite existing raster.}
}
\value{
plot or save \code{classVector} as a raster.
}
\description{
Index a class vector into a raster.
}
