% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbg8.R
\name{nbg8}
\alias{nbg8}
\title{Eight Neighbors}
\usage{
nbg8(n_row, n_col)
}
\arguments{
\item{n_row}{Integer. The number of rows of a Raster or RasterStack object.}

\item{n_col}{Integer. The number of columns of a Raster or RasterStack
object.}
}
\value{
List of length equal to the number of cells on a raster. The nth
element of the list corresponds to the 8 adjacent cell numbers of the nth
cell on the RasterStack object.
}
\description{
Return the 8 neighbors, as cell numbers, of each cell on a raster. Cell numbers
start with 1 in the upper-left corner and increase from left to right and
from top to bottom.
}
\examples{
## Matrix m mocking a raster of 3 rows and 4 columns
m <- matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
m

nbs <- nbg8(3, 4)
nbs
}
\seealso{
\code{\link[=ngbList]{ngbList()}}
}
