% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInputs.R
\name{ngbList}
\alias{ngbList}
\title{List of neighborhoods}
\usage{
ngbList(rstack, rNumb = FALSE, attTbl = NULL)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{rNumb}{logic, the neighbors of a raster cell are identified by cell
numbers if \code{rNumb=FALSE} or by row numbers if \code{rNumb=TRUE}. If
the argument is true the argument \code{attTbl} cannot be NULL.}

\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}} (see \code{\link{attTbl}}). It is required only if the
argument \code{rNumb=TRUE}.}
}
\value{
Named list of integer vectors.
}
\description{
Excluding cells on the edge of a raster, a cell with coordinates \code{(x,
y)} has 8 neighbors with coordinates: \code{(x±1, y)},  \code{(x, y±1)} and
\code{(x±1, y±1)}. The function computes the neighborhoods of all the cells
of a \code{Raster*} object. Cell with missing values are omitted.
}
\details{
\strong{Neighborhoods (\code{rNumb=FALSE})}
\itemize{
\item Neighbors are identified by their cell numbers if the argument
\code{rNumb=FALSE}.
}

\strong{Neighborhoods (\code{rNumb=TRUE})}
\itemize{
\item Neighbors are identified by their positions in the attribute table (i.e.
row numbers) if the argument \code{rNumb=TRUE} and the argument
\code{attTbl!=NULL} (see \code{\link{attTbl}});
\item When the argument \code{rNumb = TRUE}, neighbors with one or more missing
values are omitted;
\item \code{(scapes)Classifications} are faster when the list of neighborhoods
uses row numbers.
}

\strong{Neighborhood names}

The list of neighborhoods is named.
\itemize{
\item When \code{rNumb = FALSE}, the element name identify the raster cell to
which the neighborhood refers. For instance, the element with name
\code{"6"} stores the neighborhood of the raster cell \code{6}.
\item When \code{rNumb = TRUE}, the element name identify the row number to
which the neighborhood refers. For instance, the element with name
\code{"6"} stores the neighborhood of the raster cell located in the 6th
row of the attribute table (see \code{\link{attTbl}}).\cr
}
}
\note{
There is a correspondence between the indices of the attribute table
(\code{\link{attTbl}}) and the indices of the list of neighborhoods. For
instance, the first element of the list corresponds to the neighbors of the
cell stored in the first row of the attribute table.
}
\examples{
## r1, raster with cell numbers as values
r1 <- raster::raster(matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE))

## r2, raster with missing values in cell numbers 1 and 2
r2 <- raster::raster(matrix(1, nrow = 3, ncol = 4, byrow = TRUE))
r2[c(1,2)] <- NA


ngbList(raster::stack(r1,r2))
matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
}
\seealso{
\code{\link[=nbg8]{nbg8()}}, \code{\link[=attTbl]{attTbl()}}
}
