% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectFunctions.R
\name{obj.nbs}
\alias{obj.nbs}
\title{Geomorphic Units' Neighbors}
\usage{
obj.nbs(grp.bord, ngbList, only_grp = NULL, silent = FALSE)
}
\arguments{
\item{grp.bord}{named list returned by the function \code{\link{obj.border}}.
Each element of the list represents a geomorphic unit identified by the
list's names, while the values of each element of the list consist of the
indices pointing to the rows in the attribute table that correspond to the
raster cells constituting the border.}

\item{ngbList}{list, it has to contain the list of 8-neighbors of each cell
in \code{attTbl$Cell} (see \code{\link{attTbl}}). This list has to be
generated setting the argument of the function \code{\link{ngbList}}
\code{index = TRUE}. (see \code{\link{ngbList}}).}

\item{only_grp}{character vector, if geomorphic unit-IDs are provided as a
character vector, then these are the only ones considered. If this argument
is set as \code{NULL}, then all the geomorphic units included in the
argument \code{grp.bord} are considered.}

\item{silent}{logic, progress bar is not printed on the console.}
}
\value{
The function returns a named list. Each element of the list represent
a geomorphic unit as identified by the list's names. Each element is
constituted by a nested named list. The names of the nested lists are the
IDs of all the adjacent geomorphic units. The values of each nested list
consist of indices. These indices point to the rows in the attribute table
(see \code{\link{attTbl}}) that correspond to the raster cells that mark
the border between the geomorphic unit and the adjacent unit. The actual
raster cells comprising borders can be extracted from the attribute table
(e.g. \code{attTbl$Cell[indices]}) (see \code{\link{attTbl}}).
}
\description{
Identify all the units adjacent to geomorphic units of interest. For each
adjacent unit, individual borders are identified. Both function inputs and
outputs use indices pointing to raster cells in the attribute table (see
\code{\link{attTbl}} and \code{\link{ngbList}} with the argument \code{index
= TRUE}).
}
