% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piFunctions.R
\name{pi.sgm}
\alias{pi.sgm}
\title{Position index segmentation}
\usage{
pi.sgm(
  attTbl,
  ngbList,
  rNumb = FALSE,
  RO,
  mainPI,
  secPI = NULL,
  cut.mPI = NULL,
  cut.sPI = NULL,
  min.N = NULL,
  plot = FALSE,
  r = NULL
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{ngbList}{list, the list of neighborhoods returned by the function
\code{\link{ngbList}}.}

\item{rNumb}{logic, the neighborhoods of the argument \code{ngbList} are
identified by cell numbers (\code{rNumb=FALSE}) or by row numbers
(\code{rNumb=TRUE}) (see \code{\link{ngbList}}). It is advised to use row
numbers for large rasters.}

\item{RO}{column name, the name of the column with the raster object IDs.}

\item{mainPI}{column name, the name of the column with main position index
values.}

\item{secPI}{column name, the name of the column with secondary position
index values.}

\item{cut.mPI}{numeric, threshold of the main position index values below
which cells are excluded from raster objects.}

\item{cut.sPI}{numeric, threshold of the secondary position index values
below which cells are excluded from raster objects.}

\item{min.N}{numeric, the minimum number of cells a raster object has to have
to be included in the function output.}

\item{plot}{logic, plot the results.}

\item{r}{a \code{SpatRaster} object, the raster used to compute the attribute
table. Required only if \code{plot = TRUE}.}
}
\value{
The function returns a class vector with raster objects IDs. The
vector has length equal to the number of rows of the attribute table. NA
values are assigned to cells that do not belong to any raster object.
}
\description{
Segment raster objects based on position index values.
}
\details{
Raster objects are segmented based on position index values. Two
different position indices can be passed to the function (\code{mainPI} and
\code{secPI}).
\itemize{
\item Cells with Values below \code{mainPI} \strong{or} below \code{mainPI} are
excluded from raster objects;
\item Input raster objects are assigned to the same class to flag cells that
are part of a raster objects;
\item Each non-continuous group of raster object cells will identify the
an output raster objects.
}
}
\examples{
# DUMMY DATA
######################################################################################
# LOAD LIBRARIES
library(scapesClassification)
library(terra)

# LOAD THE DUMMY RASTER
r <- list.files(system.file("extdata", package = "scapesClassification"),
                pattern = "dummy_raster\\\\.tif", full.names = TRUE)
r <- terra::rast(r)

# COMPUTE THE ATTRIBUTE TABLE
at <- attTbl(r, "dummy_var")

# COMPUTE THE LIST OF NEIGBORHOODS
nbs <- ngbList(r, attTbl=at)

################################################################################
# COMPUTE RASTER OBJECTS
################################################################################
at$RO <- anchor.seed(at, nbs, silent=TRUE, class = NULL, rNumb=TRUE,
                     cond.filter = "dummy_var > 1",
                     cond.seed   = "dummy_var==max(dummy_var)",
                     cond.growth = "dummy_var<dummy_var[]",
                     lag.growth  = Inf)

# One raster object
unique(at$RO)

################################################################################
# NORMALIZED RELATIVE POSITION INDEX
################################################################################
at$relPI <- rel.pi(attTbl = at, RO = "RO", el = "dummy_var", type = "n")

################################################################################
# POSITION INDEX SEGMENTATION
################################################################################
RO1 <- pi.sgm(at, nbs,
              RO = "RO",        # Raster objects
              mainPI = "relPI", # PI segmentation layer
#'               cut.mPI = 0,      # segment on relPI values <= 0
             plot = TRUE, r = r)

graphics::layout(matrix(c(1, 2), nrow=1, byrow=TRUE))
text(xyFromCell(r,at$Cell), as.character(at$relPI)) # visualize relPI values

# Two raster object
unique(at$RO1)
}
