% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classContiguity.R
\name{reclass.nbs}
\alias{reclass.nbs}
\title{Reclassify Neighbors}
\usage{
reclass.nbs(
  attTbl,
  ngbList,
  nbsIndex = FALSE,
  classVector,
  nbs_of,
  class,
  reclass,
  reclass_all = TRUE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{ngbList}{list, it has to contain the list of 8-neighbors of each cell
in \code{attTbl$Cell} (see \code{\link{ngbList}}).}

\item{nbsIndex}{logic, \code{ngbList} contain the neighbors index position
in the attribute table (see \code{\link{ngbList}}).}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}

\item{nbs_of}{numeric or numeric vector, if a cell classified as \code{class}
is adjacent to a cell classified as indicated by the argument
\code{nbs_of}, then it is reclassified as indicated by the argument
\code{reclass}.}

\item{class}{numeric or numeric vector, indicates the classes for which
adjacent neighbors have to be evaluated.}

\item{reclass}{numeric, the numeric class to attribute to the cells meeting
contiguity conditions.}

\item{reclass_all}{logic, all cells adjacent to a reclassified cell
classified as \code{class} are reclassified as indicated by the argument
\code{reclass}.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function.
}
\description{
If members of two classes are contiguous, one of them is reclassified.
}
\details{
The function evaluates if a cell of class \code{class} is adjacent
to one of the classes of the argument \code{nbs_of} and, if \code{TRUE},
reclassifies it as indicated by the argument \code{reclass}. If the argument
\code{reclass_all} is \code{TRUE}, then all cells adjacent to a
reclassified cell classified as \code{class} are also reclassified.
}
\seealso{
\code{\link[=attTbl]{attTbl()}}, \code{\link[=ngbList]{ngbList()}}, \code{\link[=cond.reclass]{cond.reclass()}} and contiguity functions
in \code{\link[=conditions]{conditions()}}
}
