% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piFunctions.R
\name{rel.pi}
\alias{rel.pi}
\title{Standardized relative position index}
\usage{
rel.pi(attTbl, RO_col, el_col, plot = FALSE, SpatRaster = NULL)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{RO_col}{column name, the name of the attribute table column with the
raster object IDs.}

\item{el_col}{column name, the name of the attribute table column with the
elevation values on which the relative position index is computed.}

\item{plot}{logic, plot the results.}

\item{SpatRaster}{a \code{SpatRaster} object, the raster used to compute the
attribute table. Required only if \code{plot=TRUE}}
}
\value{
The function returns a vector with standardized relative position
index values. The vector has length equal to the number of rows of the
attribute table. NA values are assigned to cells that do not belong to any
raster object.
}
\description{
Compute the standardized relative position index of raster objects
}
\details{
For every raster object mean and standard deviation for elevation
values are computed. Then, the \emph{standardized relative position index} is
computed by subtracting the mean and dividing by the standard deviation for
each elevation value within raster objects.
}
\examples{
# DUMMY DATA
######################################################################################
# LOAD LIBRARIES
library(scapesClassification)
library(terra)

# LOAD THE DUMMY RASTER
r <- list.files(system.file("extdata", package = "scapesClassification"),
                pattern = "dummy_raster\\\\.tif", full.names = TRUE)
r <- terra::rast(r)

# COMPUTE THE ATTRIBUTE TABLE
at <- attTbl(r, "dummy_var")

# COMPUTE THE LIST OF NEIGBORHOODS
nbs <- ngbList(r, rNumb=TRUE, attTbl=at) # rnumb MUST be true to use obj.nbs

################################################################################
# COMPUTE RASTER OBJECTS
################################################################################
at$RO <- anchor.seed(at, nbs, silent=TRUE, class = NULL, rNumb=TRUE,
                     cond.filter = "dummy_var > 1",
                     cond.seed   = "dummy_var==max(dummy_var)",
                     cond.growth = "dummy_var<dummy_var[]",
                     lag.growth  = 0)

################################################################################
# STANDARDIZED RELATIVE POSITION INDEX
################################################################################
relPI <- rel.pi(attTbl = at, RO_col = "RO", el_col = "dummy_var",
                plot = TRUE, SpatRaster = r)

points(terra::xFromCell(r, at$Cell[which(at$RO==1)]),
       terra::yFromCell(r, at$Cell[which(at$RO==1)]) - 0.04,
       pch=20, col="red")
points(terra::xFromCell(r, at$Cell[which(at$RO==2)]),
       terra::yFromCell(r, at$Cell[which(at$RO==2)]) - 0.04,
       pch=20, col="blue")
text(xyFromCell(r,at$Cell), as.character(round(relPI,2)))
}
