% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorCells.R
\name{anchor.cell}
\alias{anchor.cell}
\title{Class vector from cell numbers}
\usage{
anchor.cell(
  attTbl,
  rstack,
  anchor,
  class,
  classVector = NULL,
  class2cell = TRUE,
  class2nbs = TRUE,
  overwrite_class = FALSE,
  plot = FALSE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}} (see \code{\link{attTbl}}).}

\item{rstack}{The \code{Raster*} object used to compute the attribute table
(see \code{\link{attTbl}}).}

\item{anchor}{integer vector, vector of raster cell numbers.}

\item{class}{numeric, the classification number to attribute to cells
indicated by the argument \code{anchor}.}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified. If provided, the function do not
attribute a class to cells that were already classified unless the argument
\code{overwrite_class = TRUE}.}

\item{class2cell}{logic, attribute the classification number to the cells
indicated by argument \code{anchor}.}

\item{class2nbs}{logic, attribute the classification number to the cells
adjacent to the ones indicated by argument \code{anchor}.}

\item{overwrite_class}{logic, reclassify cells that had already been
classified.}

\item{plot}{logic, plot \code{classVector}.}

\item{writeRaster}{filename, if a raster name is provided save the class
vector in a raster file.}

\item{overWrite}{logic, if the raster names already exist, the existing file
is overwritten.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function. If no \code{classVector} was provided, the function return a
new class vector.
}
\description{
Converts a vector of cell numbers into a class vector.
}
\details{
Converts a vector of cell numbers into a class vector. If a class
vector is provided as an input (argument \code{classVector}), then this
class vector is updated assigning a classification number to the cell
indicated by the argument \code{anchor}. The classification vector can be
indexed into the \code{rstack} using the \code{Cell} column of the
attribute table (see \code{\link{attTbl}}).
}
\examples{

\dontrun{
# LOAD LIBRARIES AND DATA
library(raster)
library(scapesClassification)

r <- list.files(system.file("extdata", package = "scapesClassification"),
                pattern = "dummy_raster\\\\.tif", full.names = TRUE)
r <- raster(r)

# SET RASTER DATA EQUAL TO CELL NUMBERS
r[]<- 1:49

# COMPUTE ATTRIBUTE TABLE AND LIST OF NEIGHBORHOODS
at  <- attTbl(r, "dummy_var")
nbs <- ngbList(r)

# CELLS FROM 1:7 IN THE `anchor` ARGUMENT
anch <-  1:7


# class2cell = TRUE & class2nbs = FALSE
cv <- anchor.cell(attTbl = at,
                  rstack = r,
                  anchor = anch,
                  class  = 10,
                  class2cell = TRUE,
                  class2nbs  = FALSE)

r_cv <- cv.2.rast(r = r, index = at$Cell, classVector = cv)

# PLOT RASTER AND CELL NUMBERS
plot(r_cv, axes=FALSE, legend = FALSE,
     colNA="#818792", col="#78b2c4", main = "class2cell=TRUE & class2nbs=FALSE")
text(r)


# class2cell = FALSE & class2nbs = TRUE
cv <- anchor.cell(attTbl = at,
                  rstack = r,
                  anchor = anch,
                  class  = 10,
                  class2cell = FALSE,
                  class2nbs  = TRUE)

r_cv <- cv.2.rast(r = r, index = at$Cell, classVector = cv)

# PLOT RASTER AND CELL NUMBERS
plot(r_cv, axes=FALSE, box=FALSE, legend = FALSE,
     colNA="#818792", col="#78b2c4", main = "class2cell=FALSE & class2nbs=TRUE")
text(r)


# class2cell = TRUE & class2nbs = TRUE
cv <- anchor.cell(attTbl = at,
                  rstack = r,
                  anchor = anch,
                  class  = 10,
                  class2cell = TRUE,
                  class2nbs  = TRUE)

r_cv <- cv.2.rast(r = r, index = at$Cell, classVector = cv)

# PLOT RASTER AND CELL NUMBERS
plot(r_cv, axes=FALSE, box=FALSE, legend = FALSE,
     colNA="#818792", col="#78b2c4", main = "class2cell=TRUE & class2nbs=TRUE")
text(r)
}
}
\seealso{
\code{\link[=anchor.seed]{anchor.seed()}}, \code{\link[=anchor.svo]{anchor.svo()}}, \code{\link[=attTbl]{attTbl()}}
}
