% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorCells.R
\name{anchor.cell}
\alias{anchor.cell}
\title{Class vector from cell numbers}
\usage{
anchor.cell(
  attTbl,
  rstack,
  anchor,
  class,
  classVector = NULL,
  class2cell = TRUE,
  class2nbs = TRUE,
  overwrite_class = FALSE,
  plot = TRUE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}} (see \code{\link{attTbl}}).}

\item{rstack}{The \code{Raster*} object used to compute the attribute table
(see \code{\link{attTbl}}).}

\item{anchor}{integer vector, vector of raster cell numbers.}

\item{class}{numeric, the classification number to attribute to cells
indicated by the argument \code{anchor}.}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified. If provided, the function do not
attribute a class to cells that were already classified unless the argument
\code{overwrite_class = TRUE}.}

\item{class2cell}{logic, attribute the classification number to the cells
indicated by argument \code{anchor}.}

\item{class2nbs}{logic, attribute the classification number to the cells
adjacent to the ones indicated by argument \code{anchor}.}

\item{overwrite_class}{logic, reclassify cells that had already been
classified.}

\item{plot}{logic, plot \code{classVector}.}

\item{writeRaster}{filename, if a raster name is provided save the class
vector in a raster file.}

\item{overWrite}{logic, if the raster names already exist, the existing file
is overwritten.}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function. If no \code{classVector} was provided, the function return a
new class vector.
}
\description{
Converts a vector of cell numbers into a class vector.
}
\details{
Converts a vector of cell numbers into a class vector. If a class
vector is provided as an input (argument \code{classVector}), then this
class vector is updated assigning a classification number to the cell
indicated by the argument \code{anchor}. The classification vector can
be indexed into the \code{rstack} using the \code{Cell} column of the
attribute table (see \code{\link{attTbl}}).
}
\seealso{
\code{\link[=anchor.seed]{anchor.seed()}}, \code{\link[=anchor.svo]{anchor.svo()}}, \code{\link[=attTbl]{attTbl()}}
}
