% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_svo.R
\name{anchor.svo}
\alias{anchor.svo}
\title{Anchor cells from spatial vector objects}
\usage{
anchor.svo(
  rstack,
  dsn,
  only_NAs = FALSE,
  fill_NAs = FALSE,
  plot = FALSE,
  saveRDS = NULL,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{dsn}{data source name (filename) or an \code{sf} or a \code{Spatial} object.}

\item{only_NAs}{logic, cell numbers extracted only for incomplete cases at
the locations of a spatial object. Incomplete cases are cells having an
NA-value in one or more layers of the \code{Raster*} object.}

\item{fill_NAs}{logic, cell numbers extracted at the locations of a spatial
object \emph{and} at contiguous locations that are incomplete cases.}

\item{plot}{logic, plot anchor points.}

\item{saveRDS}{filename, if a file name is provided save the anchor cell
vector as an RDS file.}

\item{writeRaster}{filename, if a raster name is provided save the anchor
cell vector as a raster file.}

\item{overWrite}{logic, if RDS and raster names already exist, existing files
are overwritten.}
}
\value{
Numeric vector of raster cell numbers.
}
\description{
Returns a vector of raster cell numbers extracted at the locations of a
spatial object.
}
\details{
When the arguments \code{only_NA} and \code{fill_NAs} are FALSE the
numeric output is equivalent to the output of the function
\code{raster::extract()} over a raster whose values are cell numbers.
}
\examples{
# DUMMY DATA
################################################################################
# LOAD LIBRARIES AND DATA
library(raster)
library(scapesClassification)

# CELL NUMBERS OF A DUMMY RASTER (7X7)
r_cn <- raster(matrix(1:49, nrow = 7, byrow = TRUE))

# SET SOME NA-VALUE
r_cn[c(9, 10, 11, 17, 18)] <- NA

# BULD A DUMMY POLYGON
pol <- rbind(c(0,0.95), c(0.28,1), c(0.24, 0.72), c(0.05,0.72), c(0,0.95))
pol <- spPolygons(pol)
################################################################################

################################################################################
# ANCHOR.SVO
################################################################################
ac1 <- anchor.svo(r_cn, pol, only_NAs = FALSE, fill_NAs = FALSE)
ac2 <- anchor.svo(r_cn, pol, only_NAs = TRUE,  fill_NAs = FALSE)
ac3 <- anchor.svo(r_cn, pol, only_NAs = FALSE, fill_NAs = TRUE)
ac4 <- anchor.svo(r_cn, pol, only_NAs = TRUE,  fill_NAs = TRUE)

# RASTER CELL NUMBERS 2 RASTER
r1 <- r_cn; r1[] <- NA; r1[ac1] <- 1
r2 <- r_cn; r2[] <- NA; r2[ac2] <- 1
r3 <- r_cn; r3[] <- NA; r3[ac3] <- 1
r4 <- r_cn; r4[] <- NA; r4[ac4] <- 1
################################################################################

################################################################################
# PLOTS
################################################################################
par(mfrow=c(2,2), mar=c(3, 2, 2, 2))

# 1)
plot(r1, col="#78b2c4", colNA= "grey", axes=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")
text(r_cn)
title(adj = 0.0, line = 1, sub = paste0("only_NAs = FALSE; fill_NAs = FALSE\nac=",
      paste0(sort(ac1), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))

# 2)
plot(r2, col="#78b2c4", colNA= "grey", axes=FALSE,legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")
text(r_cn)
title(adj = 0.0, line = 1, sub = paste0("only_NAs = TRUE; fill_NAs = FALSE\nac=",
      paste0(sort(ac2), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))

# 3)
plot(r3, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")
text(r_cn)
title(adj = 0.0, line = 1, sub = paste0("only_NAs = FALSE; fill_NAs = TRUE\nac=",
      paste0(sort(ac3), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))

# 4)
plot(r4, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")
text(r_cn)
title(adj = 0.0, line = 1, sub = paste0("only_NAs = TRUE; fill_NAs = TRUE\nac=",
      paste0(sort(ac4), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))
}
