% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_svo.R
\name{anchor.svo}
\alias{anchor.svo}
\title{Anchor cells from spatial vector objects}
\usage{
anchor.svo(
  rstack,
  dsn,
  only_NAs = FALSE,
  fill_NAs = FALSE,
  plot = FALSE,
  saveRDS = NULL,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{dsn}{data source name (filename) or an \code{sf} or a \code{Spatial} object.}

\item{only_NAs}{logic, cell numbers extracted only for incomplete cases at
the locations of a spatial object. Incomplete cases are cells having an
NA-value in one or more layers of the \code{Raster*} object.}

\item{fill_NAs}{logic, cell numbers extracted at the locations of a spatial
object \emph{and} at contiguous locations that are incomplete cases.}

\item{plot}{logic, plot anchor points.}

\item{saveRDS}{filename, if a file name is provided save the anchor cell
vector as an RDS file.}

\item{writeRaster}{filename, if a raster name is provided save the anchor
cell vector as a raster file.}

\item{overWrite}{logic, if RDS and raster names already exist, existing files
are overwritten.}
}
\value{
Numeric vector of raster cell numbers.
}
\description{
Returns a vector of raster cell numbers extracted at the locations of a
spatial object.
}
\details{
When the arguments \code{only_NA} and \code{fill_NAs} are FALSE the
numeric output is equivalent to the output of the function
\code{raster::extract()} over a raster whose values are cell numbers.
}
\examples{

# LOAD LIBRARIES AND DATA
library(raster)
library(scapesClassification)

# CELL NUMBERS OF A DUMMY RASTER (7X7)
r_cn <- raster(matrix(1:49, nrow = 7, byrow = TRUE))

# SET SOME NA-VALUE
r_cn[c(9, 10, 11, 17, 18)] <- NA

# BULD A DUMMY POLYGON
pol <- rbind(c(0,0.95), c(0.28,1), c(0.24, 0.72), c(0.05,0.72), c(0,0.95))
pol <- spPolygons(pol)

# EMPTY RASTER TO PLOT anchor.svo RESULTS
r2 <- r_cn; r2[] <- NA

# SET PLOT LAYOUT
par(mfrow=c(2,2), mar=c(3, 2, 2, 2))

# PLOT 1, only_NAs = FALSE; fill_NAs = FALSE
################################################################################
# ANCHOR.SVO
ac     <- anchor.svo(r_cn, pol, only_NAs = FALSE, fill_NAs = FALSE)

# ADD ANCHOR CELL TO RASTER FOR PLOTTING
r2[ac] <- 1

# PLOT
plot(r2, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")

# REFERENCE PLOT
text(r_cn)
title(adj = 0.0, line = 1,
      sub = paste0("only_NAs = FALSE; fill_NAs = FALSE\nac=",
      paste0(sort(ac), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))
################################################################################

# PLOT 2, only_NAs = TRUE; fill_NAs = FALSE
################################################################################
# ANCHOR.SVO
ac     <- anchor.svo(r_cn, pol, only_NAs = TRUE, fill_NAs = FALSE)

# ADD ANCHOR CELL TO RASTER FOR PLOTTING
r2[]   <- NA
r2[ac] <- 1

# PLOT
plot(r2, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")

# REFERENCE PLOT
text(r_cn)
title(adj = 0.0, line = 1,
      sub = paste0("only_NAs = TRUE; fill_NAs = FALSE\nac=",
      paste0(sort(ac), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))
################################################################################

# PLOT 3, only_NAs = FALSE; fill_NAs = TRUE
################################################################################
# ANCHOR.SVO
ac     <- anchor.svo(r_cn, pol, only_NAs = FALSE, fill_NAs = TRUE)

# ADD ANCHOR CELL TO RASTER FOR PLOTTING
r2[]   <- NA
r2[ac] <- 1

# PLOT
plot(r2, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")

# REFERENCE PLOT
text(r_cn)
title(adj = 0.0, line = 1,
      sub = paste0("only_NAs = FALSE; fill_NAs = TRUE\nac=",
      paste0(sort(ac), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))
################################################################################

# PLOT 4, only_NAs = TRUE; fill_NAs = TRUE
################################################################################
# ANCHOR.SVO
ac     <- anchor.svo(r_cn, pol, only_NAs = TRUE, fill_NAs = TRUE)

# ADD ANCHOR CELL TO RASTER FOR PLOTTING
r2[]   <- NA
r2[ac] <- 1

# PLOT
plot(r2, col="#78b2c4", colNA= "grey", axes=FALSE, box=FALSE, legend = FALSE, asp = NA)
plot(pol, add = TRUE, lwd = 2.5, border = "red")

# REFERENCE PLOT
text(r_cn)
title(adj = 0.0, line = 1,
      sub = paste0("only_NAs = TRUE; fill_NAs = TRUE\nac=",
      paste0(sort(ac), collapse = ",")))
legend("bottomleft", ncol = 1, bg = "white",
       legend = c("Anchor cell", "Polygon"), fill = c("#78b2c4", "red"))
}
