% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorCells.R
\name{anchor.svo}
\alias{anchor.svo}
\title{Set Anchor Cells from Spatial Vector Objects}
\usage{
anchor.svo(
  rstack,
  spatial_vector_name,
  only_NAs = FALSE,
  fill_NAs = FALSE,
  plot = TRUE,
  saveRDS = NULL,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{rstack}{\code{RasterStack} or \code{RasterLayer} object.}

\item{spatial_vector_name}{OGR data source name.}

\item{only_NAs}{logic, return only cell numbers overlapping with the spatial
vector data that have missing values in rstack.}

\item{fill_NAs}{logic, return cell numbers of cells adjacent to those
overlapping with the spatial vector data that have missing values in
rstack.}

\item{plot}{logic, plot anchor points.}

\item{saveRDS}{filename, if a file name is provided save the anchor cell
vector as an RDS file.}

\item{writeRaster}{filename, if a raster name is provided save the anchor
cell vector as a raster file.}

\item{overWrite}{logic, if RDS and raster names already exist, existing files
are overwritten.}
}
\value{
anchor cell vector.
}
\description{
Returns a vector of cell numbers overlapping with the locations of spatial
vector data. These cells can be used as anchor cells in other
\code{scapesClassification} functions.
}
\details{
When the arguments \code{only_NA} and \code{fillNAs} are FALSE the
output is equivalent to the output of the function \code{raster::extract()}
over a raster whose values are cell numbers. Cell numbers start with 1 in
the upper-left corner and increase from left to right and from top to
bottom.
}
