% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionFunctions.R
\name{cond.4.nofn}
\alias{cond.4.nofn}
\title{Test Conditions for Neighbors and Neighbors of Neighbors}
\usage{
cond.4.nofn(
  attTbl,
  ngbList,
  nbsIndex = FALSE,
  classVector,
  nbs_of,
  conditions,
  class,
  min.border = NULL,
  overwrite_class = FALSE,
  max.iter = +Inf,
  fn_perc = 1,
  directional = FALSE
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{ngbList}{list, it has to contain the list of 8-neighbors of each cell
in \code{attTbl$Cell} (see \code{\link{ngbList}}).}

\item{nbsIndex}{logic, \code{ngbList} contain the neighbors index position
in the attribute table (see \code{\link{ngbList}}).}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}

\item{nbs_of}{numeric or numeric vector, indicates the neighbors of the
classes that have to be evaluated. If it includes the class of the argument
\code{class}, the function will evaluate the conditions also for the
neighbors of the neighbors.}

\item{conditions}{character string, the conditions a cell have to meet to be
classified as indicated by the argument \code{class}. Absolute, focal cell
and focal neighborhood conditions can be used. Condition strings can
include only one neighborhood condition (see \code{\link{conditions}}).}

\item{class}{numeric, the numeric class to attribute to the cells meeting
conditions.}

\item{min.border}{numeric value between 0 and 1. It indicates the minimum
percentage of cells adjacent to the cell in evaluation that have to belong
to one of the classes in \code{nbs_of}. If one, then all neighboring cells
have to belong to one of the classes in \code{nbs_of} for the cell in
evaluation being classified as \code{class}.}

\item{overwrite_class}{logic, reclassify cells that have already been
classified.}

\item{max.iter}{integer, the maximum number of iterations.}

\item{fn_perc}{numeric value between 0 and 1. If neighborhood conditions are
considered, it determines the percentage of cells in the neighborhood for
which the conditions have to be true in order to classify the cell being
evaluated as indicted by the argument \code{class} (see
\code{\link{conditions}}).}

\item{directional}{logic, only directional neighbors are considered to test
neighborhood conditions. The argument \code{fn_perc} will also consider
only directional neighbors (see \code{\link{conditions}}).}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function.
}
\description{
Evaluate conditions for cells neighboring specific classes and classify them
if conditions are \code{TRUE}.
}
\details{
The function evaluates the conditions of the argument
\code{\link{conditions}} for all cells neighboring cells classified as one
of the classes included in \code{nbs_of}. If the argument \code{nbs_of}
includes the class of the argument \code{class}, then at each iteration the
function will evaluate if among the neighbors of the newly classified cells
there are cells meeting conditions and it will classify them accordingly.
}
\seealso{
\code{\link[=conditions]{conditions()}}, \code{\link[=attTbl]{attTbl()}}, \code{\link[=ngbList]{ngbList()}}
}
