% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionFunctions.R
\name{cond.reclass}
\alias{cond.reclass}
\title{Test Conditions and Reclassify}
\usage{
cond.reclass(
  attTbl,
  ngbList = NULL,
  rNumb = FALSE,
  classVector,
  conditions,
  class,
  reclass,
  fn_perc = 1
)
}
\arguments{
\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}.}

\item{ngbList}{list, it has to contain the list of 8-neighbors of each cell
in \code{attTbl$Cell} (see \code{\link{ngbList}}). If conditions do not
include focal neighborohood conditions this argument can be \code{NULL}
(see \code{\link{conditions}}).}

\item{rNumb}{logic, \code{ngbList} contain the neighbors index position
in the attribute table (see \code{\link{ngbList}}).}

\item{classVector}{numeric vector, defines the cells in the attribute table
that have already been classified.}

\item{conditions}{character string, the conditions a cell have to meet to be
classified as indicated by the argument \code{reclass}. Absolute and focal
neighborhood conditions can be used (see \code{\link{conditions}}).}

\item{class}{numeric or numeric vector, indicates the classes for which
conditions have to be evaluated.}

\item{reclass}{numeric, the numeric class to attribute to the cells meeting
conditions.}

\item{fn_perc}{numeric value between 0 and 1. If a focal neighborhood
condition is considered, it determines the percentage of cells in the focal
neighborhood for which the conditions have to be \code{TRUE} in order to
classify the cell being evaluated as indicted by the argument
\code{reclass} (see \code{\link{conditions}}).}
}
\value{
Update \code{classVector} with the new cells that were classified by
the function.
}
\description{
Evaluate conditions for cells of a class and reclassify them if conditions
are true.
}
\details{
The function evaluates the conditions of the argument
\code{\link{conditions}} for all cells in the classes of the argument
\code{class}. The cells for which conditions are true are reclassified as
indicted by the argument \code{reclass}.
}
\seealso{
\code{\link[=conditions]{conditions()}}, \code{\link[=attTbl]{attTbl()}}, \code{\link[=ngbList]{ngbList()}}
}
