% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\encoding{UTF-8}
\name{conditions}
\alias{conditions}
\title{scapesClassification conditions}
\usage{
conditions(names_attTbl, cond, silent = FALSE)
}
\arguments{
\item{names_attTbl}{character vector, the column (i.e. variable) names of the
attribute table returned by the function \code{\link{attTbl}}.}

\item{cond}{character string, condition string used to classify raster cells
(see \code{\link{cond.4.nofn}}, \code{\link{cond.reclass}},
\code{\link{cond.4.all}} and \code{\link{anchor.seed}}).}

\item{silent}{logic, when true, the function returns only error messages.}
}
\value{
An error message if the function finds spelling or syntax errors or a
a string with the types of rules that were detected in the argument
\code{cond}.
}
\description{
Check for spelling and syntax errors in conditions (\code{cond} argument) and
detect the type of conditions being used.
}
\details{
\strong{Conditions (or classification rules)}
\itemize{
\item Classification rules evaluate either to true or false and determine what
raster cells belong to a class.
\item Conditions are passed to \code{scapesClassification} functions as a
single character string. They can consist of combination of arithmetic
\code{(+|-|*|/|^|\%\%|\%/\%)}, relational \code{(>|<|>=|<=|==|!=|\%/\%)} and
logic operators \code{(&||)}, base R functions (e.g.,
\code{abs(variable_name)}), variables names (as named in the attribute
table, see \code{\link{attTbl}}) and previous classifications (either
stored as \code{classVector} or as rasters).
\item A combination of absolute and relative conditions can be used, but only
\emph{\strong{one neighborhood condition per string}} is allowed.
}

\cr \strong{Rule evaluation}
\itemize{
\item One of the arguments of the classification functions is the
\code{classVector}, a numeric vector that identifies what raster cells have
already been classified (non-NA values) and what have yet to be classified
(NA values). Cells that have already been classified are excluded from the
rule evaluation unless the argument \code{'overwrite_class = TRUE'}.
\item \strong{Global evaluation} \cr Classification rules are applied to all raster
cells (excluding the classified ones). This type of evaluation is common to
classification functions that do not have the argument \code{nbs_of}. Only
absolute conditions can have a global evaluation. See function
\code{\link{cond.4.all}}.
\item \strong{Focal evaluation} \cr Classification rules are applied only to raster
cells contiguous to focal cells. This type of evaluation is common to
classification functions that have the argument \code{nbs_of}. The argument
\code{nbs_of} identifies the class(es) of the focal cells. See functions
\code{\link{anchor.seed}}, \code{\link{cond.4.nofn}} and
\code{\link{cond.reclass}}.
\itemize{
\item Focal evaluation can take into account both absolute and relative rules;
\item Some classification functions do not have a \code{condition} argument.
Classifications performed by these functions are focal and take into
account only the expected spatial relationships existing among classified
and unclassified cells. See functions \code{\link{reclass.nbs}} and
\code{\link{classify.all}}.
}
}

\cr \strong{Focal evaluation, definitions}
\itemize{
\item \strong{Cell neighborhood:} a cell with coordinates \code{(x, y)} has 8
neighbors with coordinates: \code{(x±1, y)},  \code{(x, y±1)} and
\code{(x±1, y±1)}. Cells on the edge of a raster have less than 8
neighbors. See \code{\link{ngbList}}.
\item \strong{Focal cell:} cell identified by one of the classes of the argument
\code{nbs_of}.
\item \strong{Test cell:} the cell in the neighborhood of the focal cell that is
being tested. At turns all cells in the neighborhood of a focal cell are
tested against the classification rule.
\item \strong{Directional neighborhood:} it consists of the intersection between the
focal and the test cell neighborhoods.
}

\cr\strong{Absolute conditions}

\strong{1) Absolute test cell condition:} compares cell values against a
threshold value.
\itemize{
\item This type of condition applies to all functions with a \code{conditions}
argument.
\item In global evaluations all cells meeting absolute conditions receive a
classification number. In focal evaluations all \code{test cells} meeting
absolute conditions receive a classification number.
\item \emph{Examples of valid conditions:} \code{"variable_A > 1 & variable_B !=
  0"}; \code{"(variable_A^2 < 50 & variable_B == 0) | abs(variable_C) > 50"}.
\cr \emph{Functions:} \code{\link{anchor.seed}}, \code{\link{cond.4.all}},
\code{\link{cond.4.nofn}} and \code{\link{cond.reclass}}.
}

\cr \strong{2) Absolute neighborhood condition:} compares the values of the
\code{test cell} and of its \code{neighborhood} against a threshold value.
\itemize{
\item This type of condition applies to the functions \code{cond.4.nofn} and
\code{cond.reclass}.
\item An absolute neighborhood condition is identified by a variable name
followed by curly brackets (e.g., \code{"variable_name{}"}).
\item A maximum of 9 evaluations are performed for each test cell (the test
cell itself and the cells of its neighborhood are compared against a
threshold value).
\item Test cells receive a classification number if the rule is true for at
least as many evaluations as the ones specified by the argument
\code{fn_perc}. The argument \code{fn_perc} ranges from 0 to 1. When 9
evaluations are performed, \code{fn_perc = 1} means that all \code{9}
evaluations have to be true; \code{fn_perc = 0.5} means that at least
\code{4.5} (rounded to \code{5}) evaluations have to be true.
\item Only one neighborhood rule is allowed for each condition string (e.g., it
is not possible to have a condition string like \code{"variable_A{} > 0 &
  variable_B{} > 1"}).
\item The function \code{\link{cond.4.nofn}} can consider a \code{directional
  neighborhood} instead of the test cell neighborhood by setting the argument
\code{directional = TRUE}.
\item \emph{Example of valid conditions:} \code{"variable_A{} > 1 & abs(variable_B)
  != 0"}. \cr \emph{Functions:} \code{\link{cond.4.nofn}} and
\code{\link{cond.reclass}}.
}

\cr \strong{Relative conditions}

\strong{1) Relative focal cell condition:} compares the \code{test cell} value
against the \code{focal cell} value.
\itemize{
\item This type of condition applies only to functions performing focal
evaluation (i.e. function with a \code{nbs_of} argument).
\item It is identified by a variable name followed by square brackets (e.g.,
\code{"variable_name[]"}).
\item Rules are defined repeating twice the same variable name, once with
square brackets and once without. Square brackets indicate the focal cell
value. As an example, the rule \code{"dummy_var < dummy_var[]"} compares
the value of the the test cell (\code{"dummy_var"}) against the value of
the focal cell (\code{"dummy_var[]"}).
\item Test cells are classified if the rule is true.
\item \emph{Examples of valid conditions:} \code{"variable_A > variable_A[]"};
\code{"(variable_A > variable_A[] & variable_B{} < 10) | variable_C > 1"}.
Note that the last example is a combination of absolute and focal cell
conditions. \cr \emph{Functions:} \code{\link{anchor.seed}} and
\code{\link{cond.4.nofn}}.
}

\strong{2) Relative neighborhood rule:} compares the values of the \code{test
  cell} against the values of the \code{test cell neighborhood}.
\itemize{
\item This type of condition applies only to the functions
\code{\link{cond.4.nofn}} and \code{\link{cond.reclass}}.
\item It is identified by a variable name followed by curly brackets (e.g.,
\code{"variable_name{}"}).
\item Rules are defined repeating twice the same variable name, once with curly
brackets and once without. Curly brackets indicate the test cell
neighborhood. As an example, the rule \code{'dummy_var < dummy_var{}'}
compares the value of the the test cell (\code{dummy_var}) against the
values of cells included in the test cell neighborhood
(\code{dummy_var{}}).
\item A maximum of 8 evaluations are performed for each test cell (the test
cell is compared against each cell included in its neighborhood).
\item Test cells receive a classification number if the rule is true for at
least as many evaluations as the ones specified by the argument
\code{fn_perc}. The argument \code{fn_perc} ranges from 0 to 1. When 8
evaluations are performed, \code{fn_perc = 1} means that all \code{8}
evaluations have to be true; \code{fn_perc = 0.5} means that at least
\code{5} evaluations have to be true.
\item Only one neighborhood rule is allowed for each condition string (e.g., it
is not possible to have a condition string like \code{"variable_A{} > 0 &
  variable_B{} > variable_B"}).
\item The function \code{\link{cond.4.nofn}} can consider a \code{directional
  neighborhood} instead of the test cell neighborhood by setting the argument
\code{directional = TRUE}.
\item \emph{Example of valid conditions:} \code{"variable_A > variable_A{}"};
\code{"(variable_A > variable_A{} & variable_B != variable_B[]) |
  variable_C > 1"}. Note that the last example is a combination of absolute
and relative conditions. \cr \emph{Functions:} \code{\link{cond.4.nofn}} and
\code{\link{cond.reclass}}.
}
}
\examples{
# TYPES OF CONDITIONS

# As an example consider an attribute with the following columns
names_attTbl <- c("bathymetry", "slope")

# And the following conditions
cond <- "bathymetry>10"
conditions(names_attTbl, cond)

cond <- "classVector != 1"
conditions(names_attTbl, cond)

cond <- "bathymetry[]>bathymetry | abs(slope{}) < 5"
conditions(names_attTbl, cond)

cond <- "bathymetry[]>bathymetry | abs(slope{}) < slope"
conditions(names_attTbl, cond)

\dontrun{
# The function conditions detect syntax and spelling errors

cond <- "bathymetry[]>10 & | abs(slope{}) < 5"
conditions(names_attTbl, cond)

cond <- "baxxxthymetryxxx[]>10 &  abs(slope{}) < 5"
conditions(names_attTbl, cond)
}

# PREPARE PLOT
library(scapesClassification)
library(raster)
library(ggplot2)
library(reshape2)
library(ggpubr)

m     <- matrix(1:49, nrow = 7, ncol = 7, byrow = TRUE)

r   <- raster(m)
nbs <- ngbList(r)

m <- t(m)[,nrow(m):1]
m_long <- melt(m)

m_long$tags[m_long$value == 32] <- "FC"
m_long$tags[m_long$value == 25] <- "TC"

npool <- unique(unlist(nbs[c("32", "25")]))
m_long$whites[m_long$value \%in\% npool] <- m_long$value[m_long$value \%in\% npool]

m_long$tags <- as.factor(m_long$tags)

p1 <- ggplot(m_long, aes(x=Var1, y=Var2)) +
 geom_tile(aes(fill=tags), colour="gray90", lwd=1.5, show.legend = FALSE) +
 coord_fixed(ratio=1) +
 geom_text(aes(label=value), color = "grey50", family=c("serif"), size=8, na.rm=TRUE) +
 geom_text(aes(label=whites), color = "white", family=c("serif"), size=8, na.rm=TRUE) +
 scale_fill_manual(values = c("#1088a0", "goldenrod3"), na.value = "black") +
 geom_rect(aes(xmin = 2.5, xmax = 5.5, ymin = 1.5, ymax = 4.5),
 fill = "transparent", color="#1088a0", lwd = 1.5) +
 geom_rect(aes(xmin = 2.5, xmax = 5.5, ymin = 2.5, ymax = 5.5),
 fill = "transparent", color="goldenrod3", lwd = 1.5) +
 geom_rect(aes(xmin = 2.6, xmax = 5.4, ymin = 2.6, ymax = 4.4),
 fill = "transparent", color="red", lwd = 1.2) +
 theme_void()  +
 scale_x_discrete(expand=c(0,0)) + scale_y_discrete(expand=c(0,0)) +
 labs(title = "Cell numbers") +
 theme(plot.title=element_text(size = 16, face = "bold", color = "black", family = "serif"))

# FOCAL EVALUATION DEFINITIONS

p1
# FOCAL CELL
# Cell 32

# TEST CELL
# Cell 25

# FOCAL CELL NEIGHBORHOOD
# Cells 24, 25, 26, 31, 33, 38, 39, 40

# TEST CELL NEIGHBORHOOD
# Cells 17, 18, 19, 24, 26, 31, 32, 33

# DIRECTIONAL NEIGHBORHOOD
# Cells 24, 25, 26, 31, 32, 33
}
\seealso{
\code{\link[=anchor.seed]{anchor.seed()}}, \code{\link[=attTbl]{attTbl()}}, \code{\link[=cond.4.all]{cond.4.all()}}, \code{\link[=cond.4.nofn]{cond.4.nofn()}},
\code{\link[=cond.reclass]{cond.reclass()}}, \code{\link[=classify.all]{classify.all()}}
}
