% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cv.2.rast}
\alias{cv.2.rast}
\title{Class Vector To Raster}
\usage{
cv.2.rast(
  r,
  index,
  classVector,
  plot = FALSE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{r}{A \code{Raster*} object on which the values of \code{classVector}
are assigned at the raster cell numbers indicated by \code{index}. To all
remaining raster cells are assigned \code{NA values}.}

\item{index}{The cell numbers to which each element of \code{classVector}
correspond. If \code{classVector} was computed with
\code{scapesClassificaton} functions and with an attribute table
\code{\link{attTbl}}, this argument correspond to the column
\code{attTbl$Cell} of the attribute table.}

\item{classVector}{numeric vector, the values to be assigned at the raster
cell numbers indicated by \code{index}.}

\item{plot}{logic, plot anchor \code{classVector}.}

\item{writeRaster}{filename, if a raster name is provided save the
\code{classVector} as a raster file.}

\item{overWrite}{logic, overwrite existing raster.}
}
\value{
The \code{classVector} transformed into a raster.
}
\description{
Transform a class vector into a raster.
}
