% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cv.2.rast}
\alias{cv.2.rast}
\title{Class vector to raster}
\usage{
cv.2.rast(
  r,
  index,
  classVector,
  plot = FALSE,
  writeRaster = NULL,
  overWrite = FALSE
)
}
\arguments{
\item{r}{The \code{Raster*} object used to compute the attribute table (see
\code{\link{attTbl}}) or a generic \code{Raster*} object.}

\item{index}{numeric vector, the cell numbers of the argument \code{r} to
which assign the values of the argument \code{classVector}. Within
\code{scapesClassifications}, this argument normally correspond to the
column \code{Cell} of the attribute table (see \code{\link{attTbl}}).}

\item{plot}{logic, plot the raster.}

\item{writeRaster}{filename, if a raster name is provided save the raster to
a file.}

\item{overWrite}{logic, if the raster names already exist, the existing file
is overwritten.}

\item{vector}{numeric vector, the values to be assigned to the cell numbers
indicated by \code{index}.}
}
\value{
A class vector or a generic vector transformed into a raster.
}
\description{
Transform a class vector or a generic vector into a raster.
}
\details{
The arguments \code{index} and \code{vector} need to have the same
length. The function assign the values of \code{vector} at the positions of
\code{index} to an empty raster having the same spatial properties of the
raster \code{r}.
}
