% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInputs.R
\name{ngbList}
\alias{ngbList}
\title{Eight Neighbors for Complete Cases}
\usage{
ngbList(rstack, rNumb = FALSE, attTbl = NULL)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{rNumb}{logic, the 8-neighbors of each cell are identified by their row
number in the attribute table (indicated by the argument \code{attTbl}).
For instance, if cell 3 is located in row 2 it will be identified by the
number 2.}

\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}. It is required only if the argument \code{rNumb =
  TRUE}}
}
\value{
Named list of integer vectors.
}
\description{
The function \code{ngbList} returns the list of 8-neighbors for all the cells
of \code{rstack} with complete cases (i.e., cells without any missing value).
Neighbors are identified by cell numbers or by row numbers in the attribute
table (see \code{\link{attTbl}}) if the argument \code{rNumb = TRUE}. \cr\cr
The list of 8-neighbors is named. When \code{rNumb = FALSE}, names identify
the cell number for which the neighborhood was computed. When \code{rNumb =
TRUE}, names refers to row indices. For instance, the name \code{"6"} can be
used to call the neighborhood of cell number 6 when \code{rNumb = FALSE}.
However, when \code{rNumb = TRUE}, the name \code{"6"} can be used to call
the neighborhood of the raster cell stored in the 6th row of the attribute
table \code{attTbl}.\cr\cr When the argument \code{rNumb = TRUE}, only
neighbors with complete cases are considered, i.e. if a neighbor position
correspond to a raster cell with one or more missing values it is not
considered. Therefore, if a cell has all neighboring cells with missing
values, the 8-neighbor vector of that cell will have zero-length.
}
\details{
The names of the list correspond to focal cells
(\code{as.character}), the integer vectors within the list contain the
8-neighbors of each focal cell. When the argument \code{rNumb = FALSE},
both focal cells and neighbors are identified by cell numbers. When the
argument \code{rNumb = TRUE}, both focal cells and neighbors are identified
by row numbers.
}
\note{
There is a correspondence between the indices of the attribute table
(\code{\link{attTbl}}) and the indices of the 8-neighbors list. For
instance, the first element in the 8-neighbors list corresponds to the
neighbors of the cell whose values are stored in the first row of the
attribute table.
}
\examples{
## r1, raster with cell numbers as values
r1 <- raster::raster(matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE))

## r2, raster with missing values in cell numbers 1 and 2
r2 <- raster::raster(matrix(1, nrow = 3, ncol = 4, byrow = TRUE))
r2[c(1,2)] <- NA


ngbList(raster::stack(r1,r2))
matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
}
\seealso{
\code{\link[=nbg8]{nbg8()}}, \code{\link[=attTbl]{attTbl()}}
}
