% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInputs.R
\name{ngbList}
\alias{ngbList}
\title{Eight Neighbors for Complete Cases}
\usage{
ngbList(rstack, index = FALSE, attTbl = NULL)
}
\arguments{
\item{rstack}{\code{Raster*} object.}

\item{index}{Return neighbor list as index position in the attribute table.}

\item{attTbl}{data.frame, the attribute table returned by the function
\code{\link{attTbl}}. It is required only if the argument \code{index =
  TRUE}}
}
\value{
List of integer vectors.
}
\description{
Return the 8-neighbors, as cell numbers (or index), of cells on a
\code{Raster*} object with complete cases, i.e., raster cells having a value
for every layer in the stack.
}
\details{
The names of the list correspond to focal cell numbers
(\code{as.character}), the integer vectors within the list contain the
8-neighbors of each focal cell.
}
\note{
There is a correspondence between the indices of the attribute table
(\code{\link{attTbl}}) and the indices of the 8-neighbors list. For
instance, the first element in the 8-neighbors list corresponds to the
neighbors of the cell whose values are stored in the first row of the
attribute table.
}
\examples{
## r1, raster with cell numbers as values
r1 <- raster::raster(matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE))

## r2, raster with missing values in cell numbers 1 and 2
r2 <- raster::raster(matrix(1, nrow = 3, ncol = 4, byrow = TRUE))
r2[c(1,2)] <- NA


ngbList(raster::stack(r1,r2))
matrix(1:12, nrow = 3, ncol = 4, byrow = TRUE)
}
\seealso{
\code{\link[=nbg8]{nbg8()}}, \code{\link[=attTbl]{attTbl()}}
}
