% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectFunctions.R
\name{obj.border}
\alias{obj.border}
\title{Units' Border}
\usage{
obj.border(group, ngbList, silent = FALSE)
}
\arguments{
\item{group}{named list,}

\item{silent}{}
}
\value{

}
\description{
Identify the border of geomorphic units. It takes as input a list of elements each representing a single unit.
Each element can consist of the raster cell numbers composing the geomorphic unit or the indices of the raster cells
in the attribure table (see \code{\link{attTbl}})
}
\details{
elements represent the
}
